Changes according to Mentor's review
